Enum virtual_net::NetworkError

pub enum NetworkError {

    InvalidFd,
    AlreadyExists,
    Lock,
    IOError,
    AddressInUse,
    AddressNotAvailable,
    BrokenPipe,
    InsufficientMemory,
    ConnectionAborted,
    ConnectionRefused,
    ConnectionReset,
    Interrupted,
    InvalidData,
    InvalidInput,
    NotConnected,
    NoDevice,
    PermissionDenied,
    TimedOut,
    UnexpectedEof,
    WouldBlock,
    WriteZero,
    TooManyOpenFiles,
    Unsupported,
    UnknownError,

}

Variants

InvalidFd

The handle given was not usable

AlreadyExists

File exists

Lock

The filesystem has failed to lock a resource.

IOError

Something failed when doing IO. These errors can generally not be handled. It may work if tried again.

AddressInUse

The address was in use

AddressNotAvailable

The address could not be found

BrokenPipe

A pipe was closed

InsufficientMemory

Insufficient memory

ConnectionAborted

The connection was aborted

ConnectionRefused

The connection request was refused

ConnectionReset

The connection was reset

Interrupted

The operation was interrupted before it could finish

InvalidData

Invalid internal data, if the argument data is invalid, use InvalidInput

InvalidInput

The provided data is invalid

NotConnected

Could not perform the operation because there was not an open connection

NoDevice

The requested device couldnโt be accessed

PermissionDenied

Caller was not allowed to perform this operation

TimedOut

The operation did not complete within the given amount of time

UnexpectedEof

Found EOF when EOF was not expected

WouldBlock

Operation would block, this error lets the caller know that they can try again

WriteZero

A call to write returned 0

TooManyOpenFiles

Too many open files

Unsupported

The operation is not supported.

UnknownError

Some other unhandled error. If you see this, itโs probably a bug.