Module virtual_net::meta

Re-exports

pub use super::IpCidr;

pub use super::IpRoute;

pub use super::NetworkError;

pub use super::SocketStatus;

pub use super::StreamSecurity;

Structs

SocketId

Represents a socket ID

Enums

FrameSerializationFormat

MessageRequest

Message sent by the client to the server

MessageResponse

Message sent by the server back to a client

RequestType

ResponseType

Shutdown

Possible values which can be passed to the [TcpStream::shutdown] method.