Trait virtual_net::VirtualConnectionlessSocket

pub trait VirtualConnectionlessSocket: VirtualSocket + Debug + Send + Sync + 'static {
    fn poll_send_to(
        &mut self,
        cx: &mut Context<'_>,
        data: &[u8],
        addr: SocketAddr
    ) -> Poll<Result<usize>>;
    fn try_send_to(&mut self, data: &[u8], addr: SocketAddr) -> Result<usize>;
    fn poll_recv_from(
        &mut self,
        cx: &mut Context<'_>,
        buf: &mut [MaybeUninit<u8>]
    ) -> Poll<Result<(usize, SocketAddr)>>;
    fn try_recv_from(
        &mut self,
        buf: &mut [MaybeUninit<u8>]
    ) -> Result<(usize, SocketAddr)>;
}

Connectionless sockets are able to send and receive datagrams and stream bytes to multiple addresses at the same time (peer-to-peer)

Required Methods

fn poll_send_to(
&mut self,
cx: &mut Context<'_>,
data: &[u8],
addr: SocketAddr
) -> Poll<Result<usize>>

Sends out a datagram or stream of bytes on this socket to a specific address

fn try_send_to(&mut self, data: &[u8], addr: SocketAddr) -> Result<usize>

Sends out a datagram or stream of bytes on this socket to a specific address

fn poll_recv_from(
&mut self,
cx: &mut Context<'_>,
buf: &mut [MaybeUninit<u8>]
) -> Poll<Result<(usize, SocketAddr)>>

Recv a packet from the socket

fn try_recv_from(
&mut self,
buf: &mut [MaybeUninit<u8>]
) -> Result<(usize, SocketAddr)>

Recv a packet from the socket