Trait virtual_net::VirtualIoSource

pub trait VirtualIoSource: Debug + Send + Sync + 'static {
    // Required methods
    fn remove_handler(&mut self);
    fn poll_read_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>;
    fn poll_write_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>;
}

Represents an IO source

Required Methods

fn remove_handler(&mut self)

Removes a previously registered waker using a token

fn poll_read_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>

Polls the source to see if there is data waiting

fn poll_write_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>

Polls the source to see if data can be sent