Trait virtual_net::VirtualNetworking

pub trait VirtualNetworking: Debug + Send + Sync + 'static {

    fn bridge<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        network: &'life1 str,
        access_token: &'life2 str,
        security: StreamSecurity
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
    { ... }
    fn unbridge<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn dhcp_acquire<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<IpAddr>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn ip_add(&self, ip: IpAddr, prefix: u8) -> Result<()> { ... }
    fn ip_remove(&self, ip: IpAddr) -> Result<()> { ... }
    fn ip_clear(&self) -> Result<()> { ... }
    fn ip_list(&self) -> Result<Vec<IpCidr>> { ... }
    fn mac(&self) -> Result<[u8; 6]> { ... }
    fn gateway_set(&self, ip: IpAddr) -> Result<()> { ... }
    fn route_add(
        &self,
        cidr: IpCidr,
        via_router: IpAddr,
        preferred_until: Option<Duration>,
        expires_at: Option<Duration>
    ) -> Result<()> { ... }
    fn route_remove(&self, cidr: IpAddr) -> Result<()> { ... }
    fn route_clear(&self) -> Result<()> { ... }
    fn route_list(&self) -> Result<Vec<IpRoute>> { ... }
    fn bind_raw<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualRawSocket + Sync>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn listen_tcp<'life0, 'async_trait>(
        &'life0 self,
        addr: SocketAddr,
        only_v6: bool,
        reuse_port: bool,
        reuse_addr: bool
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualTcpListener + Sync>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn bind_udp<'life0, 'async_trait>(
        &'life0 self,
        addr: SocketAddr,
        reuse_port: bool,
        reuse_addr: bool
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualUdpSocket + Sync>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn bind_icmp<'life0, 'async_trait>(
        &'life0 self,
        addr: IpAddr
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualIcmpSocket + Sync>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn connect_tcp<'life0, 'async_trait>(
        &'life0 self,
        addr: SocketAddr,
        peer: SocketAddr
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualTcpSocket + Sync>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn resolve<'life0, 'life1, 'async_trait>(
        &'life0 self,
        host: &'life1 str,
        port: Option<u16>,
        dns_server: Option<IpAddr>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<IpAddr>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
    { ... }

}

An implementation of virtual networking

Provided Methods

fn bridge<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
network: &'life1 str,
access_token: &'life2 str,
security: StreamSecurity
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Bridges this local network with a remote network, which is required in order to make lower level networking calls (such as UDP/TCP)

fn unbridge<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Disconnects from the remote network essentially unbridging it

fn dhcp_acquire<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<IpAddr>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Acquires an IP address on the network and configures the routing tables

fn ip_add(&self, ip: IpAddr, prefix: u8) -> Result<()>

Adds a static IP address to the interface with a netmask prefix

fn ip_remove(&self, ip: IpAddr) -> Result<()>

Removes a static (or dynamic) IP address from the interface

fn ip_clear(&self) -> Result<()>

Clears all the assigned IP addresses for this interface

fn ip_list(&self) -> Result<Vec<IpCidr>>

Lists all the IP addresses currently assigned to this interface

fn mac(&self) -> Result<[u8; 6]>

Returns the hardware MAC address for this interface

fn gateway_set(&self, ip: IpAddr) -> Result<()>

Adds a default gateway to the routing table

fn route_add(
&self,
cidr: IpCidr,
via_router: IpAddr,
preferred_until: Option<Duration>,
expires_at: Option<Duration>
) -> Result<()>

Adds a specific route to the routing table

fn route_remove(&self, cidr: IpAddr) -> Result<()>

Removes a routing rule from the routing table

fn route_clear(&self) -> Result<()>

Clears the routing table for this interface

fn route_list(&self) -> Result<Vec<IpRoute>>

Lists all the routes defined in the routing table for this interface

fn bind_raw<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualRawSocket + Sync>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Creates a low level socket that can read and write Ethernet packets directly to the interface

fn listen_tcp<'life0, 'async_trait>(
&'life0 self,
addr: SocketAddr,
only_v6: bool,
reuse_port: bool,
reuse_addr: bool
) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualTcpListener + Sync>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Lists for TCP connections on a specific IP and Port combination Multiple servers (processes or threads) can bind to the same port if they each set the reuse-port and-or reuse-addr flags

fn bind_udp<'life0, 'async_trait>(
&'life0 self,
addr: SocketAddr,
reuse_port: bool,
reuse_addr: bool
) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualUdpSocket + Sync>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Opens a UDP socket that listens on a specific IP and Port combination Multiple servers (processes or threads) can bind to the same port if they each set the reuse-port and-or reuse-addr flags

fn bind_icmp<'life0, 'async_trait>(
&'life0 self,
addr: IpAddr
) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualIcmpSocket + Sync>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Creates a socket that can be used to send and receive ICMP packets from a paritcular IP address

fn connect_tcp<'life0, 'async_trait>(
&'life0 self,
addr: SocketAddr,
peer: SocketAddr
) -> Pin<Box<dyn Future<Output = Result<Box<dyn VirtualTcpSocket + Sync>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Opens a TCP connection to a particular destination IP address and port

fn resolve<'life0, 'life1, 'async_trait>(
&'life0 self,
host: &'life1 str,
port: Option<u16>,
dns_server: Option<IpAddr>
) -> Pin<Box<dyn Future<Output = Result<Vec<IpAddr>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Performs DNS resolution for a specific hostname