Crate virtual_net

Re-exports

pub use client::RemoteNetworkingClient;

pub use client::RemoteNetworkingClientDriver;

pub use composite::CompositeTcpListener;

pub use loopback::LoopbackNetworking;

pub use server::RemoteNetworkingServer;

pub use server::RemoteNetworkingServerDriver;

Modules

client

composite

host

loopback

meta

ruleset

rx_tx

server

tcp_pair

Structs

ArchivedIpCidr

An archived IpCidr

ArchivedIpRoute

An archived IpRoute

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

InterestGuard

InterestHandlerWaker

IpCidr

Represents an IP address and its netmask

IpCidrResolver

The resolver for an archived IpCidr

IpRoute

Represents a routing entry in the routing table of the interface

IpRouteResolver

The resolver for an archived IpRoute

UnsupportedVirtualNetworking

Enums

InterestType

NetworkError

SocketStatus

StreamSecurity

Traits

InterestHandler

VirtualConnectedSocket

Connected sockets have a persistent connection to a remote peer

VirtualConnectedSocketExt

VirtualConnectionlessSocket

Connectionless sockets are able to send and receive datagrams and stream bytes to multiple addresses at the same time (peer-to-peer)

VirtualConnectionlessSocketExt

VirtualIcmpSocket

ICMP sockets are low level devices bound to a specific address that can send and receive ICMP packets

VirtualIoSource

Represents an IO source

VirtualNetworking

An implementation of virtual networking

VirtualRawSocket

VirtualSocket

VirtualTcpListener

VirtualTcpListenerExt

VirtualTcpSocket

VirtualUdpSocket

Functions

handler_into_waker

io_err_into_net_error

net_error_into_io_err

Type Aliases

DynVirtualNetworking

Result