Crate virtual_net

Re-exports

pub use client::RemoteNetworkingClient;

pub use client::RemoteNetworkingClientDriver;

pub use server::RemoteNetworkingServer;

pub use server::RemoteNetworkingServerDriver;

Modules

client

host

meta

rx_tx

server

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

InterestGuard

InterestHandlerWaker

IpCidr

Represents an IP address and its netmask

IpRoute

Represents a routing entry in the routing table of the interface

Ipv4Addr

An IPv4 address.

Ipv6Addr

An IPv6 address.

Poller 🔒

Poller 🔒

Poller 🔒

Poller 🔒

Poller 🔒

Poller 🔒

UnsupportedVirtualNetworking

__Origin 🔒

__Origin 🔒

__Origin 🔒

__Origin 🔒

Enums

InterestType

IpAddr

An IP address, either IPv4 or IPv6.

NetworkError

SocketAddr

An internet socket address, either IPv4 or IPv6.

SocketStatus

StreamSecurity

Statics

META 🔒

__CALLSITE 🔒

Traits

InterestHandler

MustNotImplDrop 🔒

MustNotImplDrop 🔒

MustNotImplDrop 🔒

MustNotImplDrop 🔒

VirtualConnectedSocket

Connected sockets have a persistent connection to a remote peer

VirtualConnectedSocketExt

VirtualConnectionlessSocket

Connectionless sockets are able to send and receive datagrams and stream bytes to multiple addresses at the same time (peer-to-peer)

VirtualConnectionlessSocketExt

VirtualIcmpSocket

ICMP sockets are low level devices bound to a specific address that can send and receive ICMP packets

VirtualIoSource

Represents an IO source

VirtualNetworking

An implementation of virtual networking

VirtualRawSocket

VirtualSocket

VirtualTcpListener

VirtualTcpListenerExt

VirtualTcpSocket

VirtualUdpSocket

Functions

__assert_not_repr_packed 🔒

__assert_not_repr_packed 🔒

__assert_not_repr_packed 🔒

__assert_not_repr_packed 🔒

handler_into_waker

io_err_into_net_error

net_error_into_io_err

Type Definitions

DynVirtualNetworking

Result