Crate virtual_net

Modules

host

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

IpCidr

Represents an IP address and its netmask

IpRoute

Represents a routing entry in the routing table of the interface

UnsupportedVirtualNetworking

Enums

NetworkError

SocketStatus

StreamSecurity

Traits

VirtualConnectedSocket

Connected sockets have a persistent connection to a remote peer

VirtualConnectionlessSocket

Connectionless sockets are able to send and receive datagrams and stream bytes to multiple addresses at the same time (peer-to-peer)

VirtualIcmpSocket

ICMP sockets are low level devices bound to a specific address that can send and receive ICMP packets

VirtualNetworking

An implementation of virtual networking

VirtualRawSocket

VirtualSocket

VirtualTcpListener

VirtualTcpSocket

VirtualUdpSocket

Type Definitions

DynVirtualNetworking

Result