virtual_net
Module ruleset

Structs

DNSRule

Represents a DNS rule

IPV4Rule

Represents an Ipv4 rule

IPV6Rule

Represents an Ipv6 rule

RuleExpr 🔒

Ruleset

Represents a ruleset that can be used to specify a whitelist and a blacklist in order to control the inbound and outbound traffic of a network.

RulesetSegment 🔒

Enums

Direction

Represents the direction of the network traffic

DomainSpec

Specification of a domain

IPV4Spec

Specification of an Ipv4

IPV6Spec

Specification of an Ipv6 address

PortSpec

Specification of a port rule

Rule

Represents all supported rules

RuleAction 🔒

RuleParseError

Represents the errors that could happen during parsing the ruleset

RuleType 🔒

Functions

parse_as_list 🔒

parse_dns_rule 🔒

parse_enclosed 🔒

parse_ipv4_rule 🔒

parse_ipv6_rule 🔒

parse_ruleset_segments 🔒