virtual_net::server
Struct RemoteNetworkingServer

pub struct RemoteNetworkingServer {
    common: Arc<RemoteAdapterCommon>,
    inner: Arc<dyn VirtualNetworking + Send + Sync + 'static>,
}

Fields
common: Arc<RemoteAdapterCommon>inner: Arc<dyn VirtualNetworking + Send + Sync + 'static>
Implementations

impl RemoteNetworkingServer

fn new( tx: RemoteTx<MessageResponse>, rx: RemoteRx<MessageRequest>, work: UnboundedReceiver<BoxFuture<'static, ()>>, inner: Arc<dyn VirtualNetworking + Send + Sync + 'static>, ) -> (Self, RemoteNetworkingServerDriver)

pub fn new_from_mpsc( tx: Sender<MessageResponse>, rx: Receiver<MessageRequest>, inner: Arc<dyn VirtualNetworking + Send + Sync + 'static>, ) -> (Self, RemoteNetworkingServerDriver)

Creates a new interface on the remote location using a unique interface ID and a pair of channels

pub fn new_from_async_io<TX, RX>( tx: TX, rx: RX, format: FrameSerializationFormat, inner: Arc<dyn VirtualNetworking + Send + Sync + 'static>, ) -> (Self, RemoteNetworkingServerDriver)
where TX: AsyncWrite + Send + 'static, RX: AsyncRead + Send + 'static,

Creates a new interface on the remote location using a unique interface ID and a pair of channels

pub fn new_from_hyper_ws_io( tx: SplitSink<WebSocketStream<TokioIo<Upgraded>>, Message>, rx: SplitStream<WebSocketStream<TokioIo<Upgraded>>>, format: FrameSerializationFormat, inner: Arc<dyn VirtualNetworking + Send + Sync + 'static>, ) -> (Self, RemoteNetworkingServerDriver)

Creates a new interface on the remote location using a unique interface ID and a pair of channels