virtual_net
Trait VirtualConnectionlessSocket

pub trait VirtualConnectionlessSocket:
    VirtualSocket
    + Debug
    + Send
    + Sync
    + 'static {
    // Required methods
    fn try_send_to(&mut self, data: &[u8], addr: SocketAddr) -> Result<usize>;
    fn try_recv_from(
        &mut self,
        buf: &mut [MaybeUninit<u8>],
        peek: bool,
    ) -> Result<(usize, SocketAddr)>;
}

Connectionless sockets are able to send and receive datagrams and stream bytes to multiple addresses at the same time (peer-to-peer)

Required Methods

fn try_send_to(&mut self, data: &[u8], addr: SocketAddr) -> Result<usize>

Sends out a datagram or stream of bytes on this socket to a specific address

fn try_recv_from( &mut self, buf: &mut [MaybeUninit<u8>], peek: bool, ) -> Result<(usize, SocketAddr)>

Recv a packet from the socket