virtual_net

Trait VirtualIcmpSocket

pub trait VirtualIcmpSocket:
    VirtualConnectionlessSocket
    + Debug
    + Send
    + Sync
    + 'static { }

ICMP sockets are low level devices bound to a specific address that can send and receive ICMP packets